api/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [{ name = "oseiasdfarias", email = "oseias.dfarias@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic[email] (>=2.1.1,<3.0.0)",
    "fastapi (>=0.120.0,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "openai (>=2.6.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "python-dotenv (==1.0.0)",
    "redis[hiredis] (>=7.0.1,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
